feat: rename collaboration commands to live lock commands for better clarity and organization
