+++ axonotes-app/src-tauri/src/commands/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod auth_cmd;
 pub(crate) mod block_cmd;
-pub(crate) mod collaboration_cmd;
 pub(crate) mod database_cmd;
 pub(crate) mod document_cmd;
 pub(crate) mod encryption_cmd;
+pub(crate) mod live_lock_cmd;
 pub(crate) mod profile_cmd;
