+++ axonotes-app/src-tauri/src/lib.rs
@@ -49,11 +49,11 @@ pub fn run() {
             commands::document_cmd::get_document_meta,
             commands::document_cmd::list_documents,
             commands::document_cmd::update_document_metadata,
-            commands::collaboration_cmd::request_lock,
-            commands::collaboration_cmd::release_lock_focused,
-            commands::collaboration_cmd::release_lock_blur,
-            commands::collaboration_cmd::get_document_locks,
-            commands::collaboration_cmd::update_live_block,
+            commands::live_lock_cmd::request_lock,
+            commands::live_lock_cmd::release_lock_focused,
+            commands::live_lock_cmd::release_lock_blur,
+            commands::live_lock_cmd::get_document_locks,
+            commands::live_lock_cmd::update_live_block,
             commands::block_cmd::create_block,
             commands::block_cmd::get_blocks,
             commands::block_cmd::update_block,
